Split out menu utils and add tests
